arch: 5 route groups + crews/ split — codec_dashboard 3912→3618, codec_agents 1808→1128
Continuation of B6 architecture refactor (PR #160 deferrals).
Pytest: 2128 → 2148 (+20 regression tests, 0 regressions).

DASHBOARD ROUTE GROUPS (C1..C5)

Followed the routes/notifications.py pattern from PR #160 P3. Each
group is a self-contained APIRouter included in codec_dashboard.

C1 / SR-36: /api/approvals/* → routes/approvals.py
  4 endpoints (list, count, allow, deny). All state already lived in
  routes/_shared (_pending_approvals, _approval_lock,
  _evict_expired_approvals). test_bounded_dicts source-scan updated
  to look at routes/approvals.py.

C2 / SR-37: /api/heartbeat/* → routes/heartbeat.py
  4 endpoints (config GET/PUT, alerts GET/PUT). Pure CONFIG_PATH JSON
  reads, no shared mutex needed.

C3 / SR-38: /api/cortex/* → routes/cortex.py
  4 endpoints (health, skills, logs/{svc}, restart/{svc}). The skills
  endpoint reads from codec_dispatch.registry (A-4 invariant preserved).
  test_skill_loader_unification source-scan updated.

C4 / SR-39: /api/audit, /api/audit/stream, /api/audit/stats → routes/audit.py
  3 read-only endpoints. AUDIT_LOG read from routes/_shared.

C5 / SR-40: /api/observer/buffer → routes/observer.py
  1 debug-only privileged endpoint. Emits OBSERVER_BUFFER_INSPECTED on
  every read (audit trail preserved).

CREWS SPLIT (C6)

C6 / SR-41: 12 crew builders → codec_crews.py
  ~700 LOC of declarative wiring moved out of codec_agents.py.
  codec_agents now imports and re-exports the 12 builders so
  CREW_REGISTRY (which references them by name) and any external
  `from codec_agents import deep_research_crew` import keep working.
  Single-file extraction this time; future per-crew files at
  crews/{name}.py is a separate refactor.

LOC SHIFT

  codec_dashboard.py: 3,912 → 3,618 LOC (-294, -7.5%)
  codec_agents.py:    1,808 → 1,128 LOC (-680, -37.6%)

NEW TESTS

  tests/test_route_extractions.py — 20 tests across 6 phases.
  Each phase pins (a) endpoint registered on the FastAPI app via the
  router include, (b) source-side invariants (e.g. cortex_skills uses
  codec_dispatch.registry; observer emits OBSERVER_BUFFER_INSPECTED).
  Plus a global LOC floor (codec_dashboard.py < 3,700) that tightens
  as future route extractions land.

UPDATED SOURCE-SCAN TESTS

- tests/test_bounded_dicts.py: scans routes/approvals.py for
  `_evict_expired_approvals(` (was: codec_dashboard.py)
- tests/test_skill_loader_unification.py: scans routes/cortex.py for
  `from codec_dispatch import registry` (was: codec_dashboard.py)
- tests/test_security.py: scans codec_crews.py for the 12 crew
  builders + their allowed_tools= invariant (was: codec_agents.py)

Full pytest: 2148 passed, 83 skipped (env-dependent), 0 failed (1m34s).
